// src/piped-types.ts
export interface BaseStream {
  bitrate: number
  codec: string
  format: string
  indexEnd: number
  indexStart: number
  initEnd: number
  initStart: number
  mimeType: string
  quality: string
  url: string
  videoOnly: boolean
}

export interface AudioStream extends BaseStream {}

export interface VideoStream extends BaseStream {
  fps: number
  height: number
  width: number
}

export interface SearchResults<T> {
  items: T[]
  nextpage: string
  suggestion: string
  corrected: boolean
}

export interface SearchVideo {
  url: string
  type: string
  title: string
  thumbnail: string
  uploaderName: string
  uploaderUrl: string
  uploaderAvatar: string | null
  uploadedDate: string | null
  shortDescription: string | null
  duration: number
  views: number
  uploaded: number
  uploaderVerified: boolean
  isShort: boolean
}

export interface SimplifiedVideo {
  duration: number
  thumbnail: string
  title: string
  uploadedDate: string
  uplaoderAvatar: string
  uploaderUrl: string
  uploaderVerified: boolean
  url: string
  views: number
}

export interface Subtitle {
  autoGenerated: boolean
  code: string
  mimeType: string
  name: string
  url: string
}

export interface Video {
  audioStreams: AudioStream[]
  dash: null
  description: string
  dislikes: number
  duration: number
  hls: null
  lbryId: string
  likes: number
  livestream: boolean
  proxyUrl: string
  relatedStreams: SimplifiedVideo[]
  subtitles: Subtitle[]
  thumbnailUrl: string
  title: string
  uploadDate: string
  uploader: string
  uploaderUrl: string
  uploaderVerified: boolean
  videoStreams: VideoStream[]
  views: number
}